+++ src/components/edit/request-update/getRequestValues.js
@@ -1,11 +1,17 @@
 import { SCHEME_TYPES } from '../shared/index.js'
 
-const getFormIdSchemeValues = ({ requestValues }) => {
-    const idSchemeProps = [
+const getFormIdSchemeValues = ({
+    requestValues,
+    outputDataItemIdSchemeAvailable,
+}) => {
+    const baseIdSchemeProps = [
         'outputIdScheme',
         'outputDataElementIdScheme',
         'outputOrgUnitIdScheme',
     ]
+    const idSchemeProps = outputDataItemIdSchemeAvailable
+        ? [...baseIdSchemeProps, 'outputDataItemIdScheme']
+        : baseIdSchemeProps
     return idSchemeProps.reduce((idSchemeValues, prop) => {
         const attributeProp = `source_${prop}_attribute`
 
@@ -18,7 +24,10 @@ const getFormIdSchemeValues = ({ requestValues }) => {
     }, {})
 }
 
-export const getRequestValuesFromForm = ({ requestValues }) => {
+export const getRequestValuesFromForm = ({
+    requestValues,
+    outputDataItemIdSchemeAvailable,
+}) => {
     const validFilters = !requestValues.filtersUsed
         ? []
         : requestValues?.filtersInfo
@@ -46,17 +55,23 @@ export const getRequestValuesFromForm = ({ requestValues }) => {
             ? null
             : requestValues.visualizationInfo,
         inputIdScheme: SCHEME_TYPES.uid,
-        ...getFormIdSchemeValues({ requestValues }),
+        ...getFormIdSchemeValues({
+            requestValues,
+            outputDataItemIdSchemeAvailable,
+        }),
     }
 }
 
-const getIdSchemeValues = ({ request }) => {
+const getIdSchemeValues = ({ request, outputDataItemIdSchemeAvailable }) => {
     const defaultSchemeProp = 'outputIdScheme'
-    const idSchemeProps = [
+    const baseIdSchemeProps = [
         defaultSchemeProp,
         'outputDataElementIdScheme',
         'outputOrgUnitIdScheme',
     ]
+    const idSchemeProps = outputDataItemIdSchemeAvailable
+        ? [...baseIdSchemeProps, 'outputDataItemIdScheme']
+        : baseIdSchemeProps
     return idSchemeProps.reduce((idSchemeValues, prop) => {
         idSchemeValues[`source_${prop}`] = request?.[prop]
             ? request?.[prop]?.split(':')[0]?.toUpperCase()
@@ -69,7 +84,10 @@ const getIdSchemeValues = ({ request }) => {
     }, {})
 }
 
-export const getInitialValuesFromRequest = ({ request }) => ({
+export const getInitialValuesFromRequest = ({
+    request,
+    outputDataItemIdSchemeAvailable,
+}) => ({
     requestName: request?.name,
     peInfo: request?.peInfo ?? [],
     ouInfo: request?.ouInfo ?? [],
@@ -78,5 +96,5 @@ export const getInitialValuesFromRequest = ({ request }) => ({
     filtersInfo: request?.filtersInfo ?? [{ dimension: null }],
     visualizationLinked: Boolean(request.visualization),
     visualizationInfo: request?.visualizationInfo ?? null,
-    ...getIdSchemeValues({ request }),
+    ...getIdSchemeValues({ request, outputDataItemIdSchemeAvailable }),
 })